+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 75cf018 - 2017-07-24: Added code to properly convert seeder and leechers into integers 
+- * 891dc86 - 2017-07-24: Added code to properly convert seeder and leechers into integers 
+- * dc71cbc - 2017-07-24: Added Nextorrent torrent provider 
 - * 45170ac - 2017-07-23: Fixed Unicode issues with general config page Misc clean up of code 
 - * 02a28f7 - 2017-07-22: Fixed issues with provider url&#x27;s Fixed issues with downloading torrents via uTorrent 
 - * 22ffaa3 - 2017-07-22: Fixed torrentday url